Change p2lab to run with config file
Example config.yaml added to the root of the repo.

Dockerfile now contains a new environment variable which tells the container where it should look for the config file.

Additional dependency on pyyaml to parse the config file.

__main__.py now parses the arguments in from the config file rather than from inputs.

README.md updated with the command that runs the container with a custom config file. By default the docker image will be build with the config.yaml file from the repository. This can be overridden by mounting a config from the local filesystem using the -v flag as shown in the README.md example.